Stable login system with roles

+++ app/Http/Controllers/Admin/UsersController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Admin;
 
 use App\User;
 use App\Role;
+use Gate;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 
@@ -29,11 +30,14 @@ class UsersController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  \App\User  $user
+     * @param \App\User $user
      * @return \Illuminate\Http\Response
      */
     public function edit(User $user)
     {
+        if (Gate::denies('edit-users')) {
+            return redirect(route('admin.users.index'));
+        }
         $roles = Role::all();
         return view('admin.users.edit')->with([
             'user' => $user,
@@ -44,24 +48,39 @@ class UsersController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\User  $user
+     * @param \Illuminate\Http\Request $request
+     * @param \App\User $user
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, User $user)
     {
         $user->roles()->sync($request->roles);
+        $user->name = $request->name;
+        $user->email = $request->email;
+        if($user->save()){
+            $request->session()->flash('success', $user->name. ' Güncelleme işlemi başarılı.');
+        }else{
+            $request->session()->flash('error','Bir şeyler ters gitti.');
+        }
         return redirect()->route('admin.users.index');
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\User  $user
+     * @param \App\User $user
      * @return \Illuminate\Http\Response
      */
     public function destroy(User $user)
     {
+        if (Gate::denies('delete-users')) {
+            ?>
+            <div class="alert alert-danger" role="alert">
+                This is a danger alert—check it out!
+            </div>
+            <?php
+            return redirect(route('admin.users.index'));
+        }
         $user->roles()->detach();
         $user->delete();
         return redirect()->route('admin.users.index');
